web-app/requirements.txt
Flask == 3.1.0
python-dotenv == 1.0.0
numpy == 2.2.4
requests == 2.31.0
pytest